refactor(core,test): rename namespaces and cleanup using directives
