+++ Test/EntityCollectorTestUnit.cs
@@ -1,10 +1,6 @@
-using System.Collections.Generic;
-using CoreECS;
 using CoreECS.Defines;
-using NUnit.Framework;
-using TinyECS;
 
-namespace TinyECS.Test
+namespace CoreECS.Test
 {
     [TestFixture]
     public class EntityCollectorTestUnit
