+++ Test/EntityManagerTestUnit.cs
@@ -1,12 +1,7 @@
-using System;
-using System.Collections.Generic;
-using CoreECS;
 using CoreECS.Defines;
 using CoreECS.Managers;
-using NUnit.Framework;
-using TinyECS;
 
-namespace TinyECS.Test
+namespace CoreECS.Test
 {
     [TestFixture]
     public class EntityManagerTestUnit
