+++ Test/EntityTestUnit.cs
@@ -1,11 +1,6 @@
-using System;
-using System.Collections.Generic;
-using CoreECS;
 using CoreECS.Defines;
-using NUnit.Framework;
-using TinyECS;
 
-namespace TinyECS.Test
+namespace CoreECS.Test
 {
     [TestFixture]
     public class EntityTestUnit
