+++ Test/SystemTestUnit.cs
@@ -1,11 +1,7 @@
-using System;
-using CoreECS;
 using CoreECS.Defines;
 using CoreECS.Managers;
-using NUnit.Framework;
-using TinyECS;
 
-namespace TinyECS.Test
+namespace CoreECS.Test
 {
     [TestFixture]
     public class SystemTestUnit
